refactor(MainScreen): clean code
- Add alarm animation

+++ app/src/main/java/com/wem/snoozy/data/local/AlarmItemModel.kt
@@ -9,7 +9,7 @@ data class AlarmItemModel(
     @PrimaryKey(autoGenerate = true) val id: Int,
     val ringDay: String,
     val ringHours: String,
-    val ringHoursMillis: Long,
+    val ringHoursMillis: Int,
     val timeToBed: String,
     val checked: Boolean,
     val repeatDays: String
